+++ IEC/src/Domain/Enums/MovieRoleEnum.cs
@@ -1,6 +1,6 @@
 namespace Domain.Enums
 {
-    public enum MovieRoleEnum : byte
+    public enum MovieRoleEnum : int
     {
         Star = 1,
         Director = 2
